+++ usecase/user_test.go
@@ -65,43 +65,3 @@ func TestUsecaseDeleteUser(t *testing.T) {
 		t.Error(err)
 	}
 }
-
-func TestGenerateToken(t *testing.T) {
-	tokenPersistence := persistence.NewTokenPersistence()
-	tokenUseCase := NewTokenUseCase(tokenPersistence)
-
-	_, err := tokenUseCase.GenerateToken(testUser.UserID)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-func TestValidateToken(t *testing.T) {
-	tokenPersistence := persistence.NewTokenPersistence()
-	tokenUseCase := NewTokenUseCase(tokenPersistence)
-
-	token, err := tokenUseCase.GenerateToken(testUser.UserID)
-	if err != nil {
-		t.Error(err)
-	}
-
-	userID, err := tokenUseCase.ValidateToken(token)
-	if err != nil {
-		t.Error(err)
-	}
-	if userID != testUser.UserID {
-		t.Errorf("UserID is not correct")
-	}
-}
-
-func TestValidateInvalidToken(t *testing.T) {
-	tokenPersistence := persistence.NewTokenPersistence()
-	tokenUseCase := NewTokenUseCase(tokenPersistence)
-
-	token := "invalid_token"
-
-	_, err := tokenUseCase.ValidateToken(token)
-	if err == nil {
-		t.Errorf("Invalid token is validated")
-	}
-}
